+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+## Changed
+- converting oform to pdf (since document server v8.0)
+
 ## Added
 - disable plugins
 - macro launch